+++ src/ui/map.test.ts
@@ -76,41 +76,6 @@ describe('Map', () => {
         //t.error();
     });
 
-    describe('disables handlers', () => {
-        test('disables all handlers', () => {
-            const map = createMap({interactive: false});
-
-            expect(map.boxZoom.isEnabled()).toBeFalsy();
-            expect(map.doubleClickZoom.isEnabled()).toBeFalsy();
-            expect(map.dragPan.isEnabled()).toBeFalsy();
-            expect(map.dragRotate.isEnabled()).toBeFalsy();
-            expect(map.keyboard.isEnabled()).toBeFalsy();
-            expect(map.scrollZoom.isEnabled()).toBeFalsy();
-            expect(map.touchZoomRotate.isEnabled()).toBeFalsy();
-        });
-
-        const handlerNames = [
-            'scrollZoom',
-            'boxZoom',
-            'dragRotate',
-            'dragPan',
-            'keyboard',
-            'doubleClickZoom',
-            'touchZoomRotate'
-        ];
-        handlerNames.forEach((handlerName) => {
-            test(`disables "${handlerName}" handler`, () => {
-                const options = {};
-                options[handlerName] = false;
-                const map = createMap(options);
-
-                expect(map[handlerName].isEnabled()).toBeFalsy();
-
-            });
-        });
-
-    });
-
     test('emits load event after a style is set', done => {
         const map = new Map({container: window.document.createElement('div')} as any as MapOptions);
 
@@ -211,251 +176,6 @@ describe('Map', () => {
         });
     });
 
-    test('#moveLayer', async () => {
-        const map = createMap({
-            style: extend(createStyle(), {
-                sources: {
-                    mapbox: {
-                        type: 'vector',
-                        minzoom: 1,
-                        maxzoom: 10,
-                        tiles: ['http://example.com/{z}/{x}/{y}.png']
-                    }
-                },
-                layers: [{
-                    id: 'layerId1',
-                    type: 'circle',
-                    source: 'mapbox',
-                    'source-layer': 'sourceLayer'
-                }, {
-                    id: 'layerId2',
-                    type: 'circle',
-                    source: 'mapbox',
-                    'source-layer': 'sourceLayer'
-                }]
-            })
-        });
-
-        await map.once('render');
-        map.moveLayer('layerId1', 'layerId2');
-        expect(map.getLayer('layerId1').id).toBe('layerId1');
-        expect(map.getLayer('layerId2').id).toBe('layerId2');
-    });
-
-    test('#getLayer', async () => {
-        const layer = {
-            id: 'layerId',
-            type: 'circle',
-            source: 'mapbox',
-            'source-layer': 'sourceLayer'
-        };
-        const map = createMap({
-            style: extend(createStyle(), {
-                sources: {
-                    mapbox: {
-                        type: 'vector',
-                        minzoom: 1,
-                        maxzoom: 10,
-                        tiles: ['http://example.com/{z}/{x}/{y}.png']
-                    }
-                },
-                layers: [layer]
-            })
-        });
-
-        await map.once('render');
-        const mapLayer = map.getLayer('layerId');
-        expect(mapLayer.id).toBe(layer.id);
-        expect(mapLayer.type).toBe(layer.type);
-        expect(mapLayer.source).toBe(layer.source);
-    });
-
-    describe('#getLayersOrder', () => {
-        test('returns ids of layers in the correct order', done => {
-            const map = createMap({
-                style: extend(createStyle(), {
-                    'sources': {
-                        'raster': {
-                            type: 'raster',
-                            tiles: ['http://tiles.server']
-                        }
-                    },
-                    'layers': [{
-                        'id': 'raster',
-                        'type': 'raster',
-                        'source': 'raster'
-                    }]
-                })
-            });
-
-            map.on('style.load', () => {
-                map.addLayer({
-                    id: 'custom',
-                    type: 'custom',
-                    render() {}
-                }, 'raster');
-                expect(map.getLayersOrder()).toEqual(['custom', 'raster']);
-                done();
-            });
-        });
-    });
-
-    test('#setMinZoom', () => {
-        const map = createMap({zoom: 5});
-        map.setMinZoom(3.5);
-        map.setZoom(1);
-        expect(map.getZoom()).toBe(3.5);
-    });
-
-    test('unset minZoom', () => {
-        const map = createMap({minZoom: 5});
-        map.setMinZoom(null);
-        map.setZoom(1);
-        expect(map.getZoom()).toBe(1);
-    });
-
-    test('#getMinZoom', () => {
-        const map = createMap({zoom: 0});
-        expect(map.getMinZoom()).toBe(-2);
-        map.setMinZoom(10);
-        expect(map.getMinZoom()).toBe(10);
-    });
-
-    test('ignore minZooms over maxZoom', () => {
-        const map = createMap({zoom: 2, maxZoom: 5});
-        expect(() => {
-            map.setMinZoom(6);
-        }).toThrow();
-        map.setZoom(0);
-        expect(map.getZoom()).toBe(0);
-    });
-
-    test('#setMaxZoom', () => {
-        const map = createMap({zoom: 0});
-        map.setMaxZoom(3.5);
-        map.setZoom(4);
-        expect(map.getZoom()).toBe(3.5);
-    });
-
-    test('unset maxZoom', () => {
-        const map = createMap({maxZoom: 5});
-        map.setMaxZoom(null);
-        map.setZoom(6);
-        expect(map.getZoom()).toBe(6);
-    });
-
-    test('#getMaxZoom', () => {
-        const map = createMap({zoom: 0});
-        expect(map.getMaxZoom()).toBe(22);
-        map.setMaxZoom(10);
-        expect(map.getMaxZoom()).toBe(10);
-    });
-
-    test('ignore maxZooms over minZoom', () => {
-        const map = createMap({minZoom: 5});
-        expect(() => {
-            map.setMaxZoom(4);
-        }).toThrow();
-        map.setZoom(5);
-        expect(map.getZoom()).toBe(5);
-    });
-
-    test('throw on maxZoom smaller than minZoom at init', () => {
-        expect(() => {
-            createMap({minZoom: 10, maxZoom: 5});
-        }).toThrow(new Error('maxZoom must be greater than or equal to minZoom'));
-    });
-
-    test('throw on maxZoom smaller than minZoom at init with falsey maxZoom', () => {
-        expect(() => {
-            createMap({minZoom: 1, maxZoom: 0});
-        }).toThrow(new Error('maxZoom must be greater than or equal to minZoom'));
-    });
-
-    test('#setMinPitch', () => {
-        const map = createMap({pitch: 20});
-        map.setMinPitch(10);
-        map.setPitch(0);
-        expect(map.getPitch()).toBe(10);
-    });
-
-    test('unset minPitch', () => {
-        const map = createMap({minPitch: 20});
-        map.setMinPitch(null);
-        map.setPitch(0);
-        expect(map.getPitch()).toBe(0);
-    });
-
-    test('#getMinPitch', () => {
-        const map = createMap({pitch: 0});
-        expect(map.getMinPitch()).toBe(0);
-        map.setMinPitch(10);
-        expect(map.getMinPitch()).toBe(10);
-    });
-
-    test('ignore minPitchs over maxPitch', () => {
-        const map = createMap({pitch: 0, maxPitch: 10});
-        expect(() => {
-            map.setMinPitch(20);
-        }).toThrow();
-        map.setPitch(0);
-        expect(map.getPitch()).toBe(0);
-    });
-
-    test('#setMaxPitch', () => {
-        const map = createMap({pitch: 0});
-        map.setMaxPitch(10);
-        map.setPitch(20);
-        expect(map.getPitch()).toBe(10);
-    });
-
-    test('unset maxPitch', () => {
-        const map = createMap({maxPitch: 10});
-        map.setMaxPitch(null);
-        map.setPitch(20);
-        expect(map.getPitch()).toBe(20);
-    });
-
-    test('#getMaxPitch', () => {
-        const map = createMap({pitch: 0});
-        expect(map.getMaxPitch()).toBe(60);
-        map.setMaxPitch(10);
-        expect(map.getMaxPitch()).toBe(10);
-    });
-
-    test('ignore maxPitchs over minPitch', () => {
-        const map = createMap({minPitch: 10});
-        expect(() => {
-            map.setMaxPitch(0);
-        }).toThrow();
-        map.setPitch(10);
-        expect(map.getPitch()).toBe(10);
-    });
-
-    test('throw on maxPitch smaller than minPitch at init', () => {
-        expect(() => {
-            createMap({minPitch: 10, maxPitch: 5});
-        }).toThrow(new Error('maxPitch must be greater than or equal to minPitch'));
-    });
-
-    test('throw on maxPitch smaller than minPitch at init with falsey maxPitch', () => {
-        expect(() => {
-            createMap({minPitch: 1, maxPitch: 0});
-        }).toThrow(new Error('maxPitch must be greater than or equal to minPitch'));
-    });
-
-    test('throw on maxPitch greater than valid maxPitch at init', () => {
-        expect(() => {
-            createMap({maxPitch: 90});
-        }).toThrow(new Error('maxPitch must be less than or equal to 85'));
-    });
-
-    test('throw on minPitch less than valid minPitch at init', () => {
-        expect(() => {
-            createMap({minPitch: -10});
-        }).toThrow(new Error('minPitch must be greater than or equal to 0'));
-    });
-
     test('#remove', () => {
         const map = createMap();
         const spyWorkerPoolRelease = jest.spyOn(map.style.dispatcher.workerPool, 'release');
@@ -503,33 +223,6 @@ describe('Map', () => {
             expect(onRemoveCalled).toBe(1);
             done();
         });
-    });
-
-    test('does not fire "webglcontextlost" after #remove has been called', done => {
-        const map = createMap();
-        const canvas = map.getCanvas();
-        map.once('webglcontextlost', () => done('"webglcontextlost" fired after #remove has been called'));
-        map.remove();
-        // Dispatch the event manually because at the time of this writing, gl does not support
-        // the WEBGL_lose_context extension.
-        canvas.dispatchEvent(new window.Event('webglcontextlost'));
-        done();
-    });
-
-    test('does not fire "webglcontextrestored" after #remove has been called', done => {
-        const map = createMap();
-        const canvas = map.getCanvas();
-
-        map.once('webglcontextlost', () => {
-            map.once('webglcontextrestored', () => done('"webglcontextrestored" fired after #remove has been called'));
-            map.remove();
-            canvas.dispatchEvent(new window.Event('webglcontextrestored'));
-            done();
-        });
-
-        // Dispatch the event manually because at the time of this writing, gl does not support
-        // the WEBGL_lose_context extension.
-        canvas.dispatchEvent(new window.Event('webglcontextlost'));
     });
 
     test('#remove broadcasts removeMap to worker', () => {
@@ -611,28 +304,6 @@ describe('Map', () => {
     test('#unproject', () => {
         const map = createMap();
         expect(fixedLngLat(map.unproject([100, 100]))).toEqual({lng: 0, lat: 0});
-    });
-
-    test('#listImages', done => {
-        const map = createMap();
-
-        map.on('load', () => {
-            expect(map.listImages()).toHaveLength(0);
-
-            map.addImage('img', {width: 1, height: 1, data: new Uint8Array(4)});
-
-            const images = map.listImages();
-            expect(images).toHaveLength(1);
-            expect(images[0]).toBe('img');
-            done();
-        });
-    });
-
-    test('#listImages throws an error if called before "load"', () => {
-        const map = createMap();
-        expect(() => {
-            map.listImages();
-        }).toThrow(Error);
     });
 
     describe('#queryRenderedFeatures', () => {
@@ -717,301 +388,6 @@ describe('Map', () => {
         test('returns an empty array when no style is loaded', () => {
             const map = createMap({style: undefined});
             expect(map.queryRenderedFeatures()).toEqual([]);
-        });
-
-    });
-
-    describe('#setLayoutProperty', () => {
-        test('sets property', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'geojson': {
-                            'type': 'geojson',
-                            'data': {
-                                'type': 'FeatureCollection',
-                                'features': []
-                            }
-                        }
-                    },
-                    'layers': [{
-                        'id': 'symbol',
-                        'type': 'symbol',
-                        'source': 'geojson',
-                        'layout': {
-                            'text-transform': 'uppercase'
-                        }
-                    }]
-                }
-            });
-
-            map.on('style.load', () => {
-                map.style.dispatcher.broadcast = function (key, value: any) {
-                    expect(key).toBe(MessageType.updateLayers);
-                    expect(value.layers.map((layer) => { return layer.id; })).toEqual(['symbol']);
-                    return Promise.resolve({} as any);
-                };
-
-                map.setLayoutProperty('symbol', 'text-transform', 'lowercase');
-                map.style.update({} as EvaluationParameters);
-                expect(map.getLayoutProperty('symbol', 'text-transform')).toBe('lowercase');
-                done();
-            });
-        });
-
-        test('throw before loaded', () => {
-            const map = createMap({
-                style: {
-                    version: 8,
-                    sources: {},
-                    layers: []
-                }
-            });
-
-            expect(() => {
-                map.setLayoutProperty('symbol', 'text-transform', 'lowercase');
-            }).toThrow(Error);
-
-        });
-
-        test('fires an error if layer not found', done => {
-            const map = createMap({
-                style: {
-                    version: 8,
-                    sources: {},
-                    layers: []
-                }
-            });
-
-            map.on('style.load', () => {
-                map.on('error', ({error}) => {
-                    expect(error.message).toMatch(/Cannot style non-existing layer "non-existant"./);
-                    done();
-                });
-                map.setLayoutProperty('non-existant', 'text-transform', 'lowercase');
-            });
-        });
-
-        test('fires a data event', async () => {
-            // background layers do not have a source
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {},
-                    'layers': [{
-                        'id': 'background',
-                        'type': 'background',
-                        'layout': {
-                            'visibility': 'none'
-                        }
-                    }]
-                }
-            });
-
-            await map.once('style.load');
-            const dataPromise = map.once('data');
-            map.setLayoutProperty('background', 'visibility', 'visible');
-            const e = await dataPromise;
-            expect(e.dataType).toBe('style');
-        });
-
-        test('sets visibility on background layer', done => {
-            // background layers do not have a source
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {},
-                    'layers': [{
-                        'id': 'background',
-                        'type': 'background',
-                        'layout': {
-                            'visibility': 'none'
-                        }
-                    }]
-                }
-            });
-
-            map.on('style.load', () => {
-                map.setLayoutProperty('background', 'visibility', 'visible');
-                expect(map.getLayoutProperty('background', 'visibility')).toBe('visible');
-                done();
-            });
-        });
-
-        test('sets visibility on raster layer', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'maplibre-satellite': {
-                            'type': 'raster',
-                            'tiles': ['http://example.com/{z}/{x}/{y}.png']
-                        }
-                    },
-                    'layers': [{
-                        'id': 'satellite',
-                        'type': 'raster',
-                        'source': 'maplibre-satellite',
-                        'layout': {
-                            'visibility': 'none'
-                        }
-                    }]
-                }
-            });
-
-            // Suppress errors because we're not loading tiles from a real URL.
-            map.on('error', () => {});
-
-            map.on('style.load', () => {
-                map.setLayoutProperty('satellite', 'visibility', 'visible');
-                expect(map.getLayoutProperty('satellite', 'visibility')).toBe('visible');
-                done();
-            });
-        });
-
-        test('sets visibility on video layer', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'drone': {
-                            'type': 'video',
-                            'urls': [],
-                            'coordinates': [
-                                [-122.51596391201019, 37.56238816766053],
-                                [-122.51467645168304, 37.56410183312965],
-                                [-122.51309394836426, 37.563391708549425],
-                                [-122.51423120498657, 37.56161849366671]
-                            ]
-                        }
-                    },
-                    'layers': [{
-                        'id': 'shore',
-                        'type': 'raster',
-                        'source': 'drone',
-                        'layout': {
-                            'visibility': 'none'
-                        }
-                    }]
-                }
-            });
-
-            map.on('style.load', () => {
-                map.setLayoutProperty('shore', 'visibility', 'visible');
-                expect(map.getLayoutProperty('shore', 'visibility')).toBe('visible');
-                done();
-            });
-        });
-
-        test('sets visibility on image layer', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {
-                        'image': {
-                            'type': 'image',
-                            'url': '',
-                            'coordinates': [
-                                [-122.51596391201019, 37.56238816766053],
-                                [-122.51467645168304, 37.56410183312965],
-                                [-122.51309394836426, 37.563391708549425],
-                                [-122.51423120498657, 37.56161849366671]
-                            ]
-                        }
-                    },
-                    'layers': [{
-                        'id': 'image',
-                        'type': 'raster',
-                        'source': 'image',
-                        'layout': {
-                            'visibility': 'none'
-                        }
-                    }]
-                }
-            });
-
-            map.on('style.load', () => {
-                map.setLayoutProperty('image', 'visibility', 'visible');
-                expect(map.getLayoutProperty('image', 'visibility')).toBe('visible');
-                done();
-            });
-        });
-
-    });
-
-    describe('#getLayoutProperty', () => {
-        test('fires an error if layer not found', done => {
-            const map = createMap({
-                style: {
-                    version: 8,
-                    sources: {},
-                    layers: []
-                }
-            });
-
-            map.on('style.load', () => {
-                map.on('error', ({error}) => {
-                    expect(error.message).toMatch(/Cannot get style of non-existing layer "non-existant"./);
-                    done();
-                });
-                (map as any).getLayoutProperty('non-existant', 'text-transform', 'lowercase');
-            });
-        });
-
-    });
-
-    describe('#setPaintProperty', () => {
-        test('sets property', done => {
-            const map = createMap({
-                style: {
-                    'version': 8,
-                    'sources': {},
-                    'layers': [{
-                        'id': 'background',
-                        'type': 'background'
-                    }]
-                }
-            });
-
-            map.on('style.load', () => {
-                map.setPaintProperty('background', 'background-color', 'red');
-                expect(map.getPaintProperty('background', 'background-color')).toBe('red');
-                done();
-            });
-        });
-
-        test('throw before loaded', () => {
-            const map = createMap({
-                style: {
-                    version: 8,
-                    sources: {},
-                    layers: []
-                }
-            });
-
-            expect(() => {
-                map.setPaintProperty('background', 'background-color', 'red');
-            }).toThrow(Error);
-
-        });
-
-        test('fires an error if layer not found', done => {
-            const map = createMap({
-                style: {
-                    version: 8,
-                    sources: {},
-                    layers: []
-                }
-            });
-
-            map.on('style.load', () => {
-                map.on('error', ({error}) => {
-                    expect(error.message).toMatch(/Cannot style non-existing layer "non-existant"./);
-                    done();
-                });
-                map.setPaintProperty('non-existant', 'background-color', 'red');
-            });
         });
 
     });
@@ -1117,37 +493,6 @@ describe('Map', () => {
         expect(map.isMoving()).toBeFalsy();
     });
 
-    test('#removeLayer restores Map#loaded() to true', done => {
-        const map = createMap({
-            style: extend(createStyle(), {
-                sources: {
-                    mapbox: {
-                        type: 'vector',
-                        minzoom: 1,
-                        maxzoom: 10,
-                        tiles: ['http://example.com/{z}/{x}/{y}.png']
-                    }
-                },
-                layers: [{
-                    id: 'layerId',
-                    type: 'circle',
-                    source: 'mapbox',
-                    'source-layer': 'sourceLayer'
-                }]
-            })
-        });
-
-        map.once('render', () => {
-            map.removeLayer('layerId');
-            map.on('render', () => {
-                if (map.loaded()) {
-                    map.remove();
-                    done();
-                }
-            });
-        });
-    });
-
     test('stops camera animation on mousedown when interactive', () => {
         const map = createMap({interactive: true});
         map.flyTo({center: [200, 0], duration: 100});
@@ -1202,260 +547,11 @@ describe('Map', () => {
 
     });
 
-    test('map fires `styleimagemissing` for missing icons', async () => {
-        const map = createMap();
-
-        const id = 'missing-image';
-
-        const sampleImage = {width: 2, height: 1, data: new Uint8Array(8)};
-
-        let called: string;
-        map.on('styleimagemissing', e => {
-            map.addImage(e.id, sampleImage);
-            called = e.id;
-        });
-
-        expect(map.hasImage(id)).toBeFalsy();
-
-        const generatedImage = await map.style.imageManager.getImages([id]);
-        expect(generatedImage[id].data.width).toEqual(sampleImage.width);
-        expect(generatedImage[id].data.height).toEqual(sampleImage.height);
-        expect(generatedImage[id].data.data).toEqual(sampleImage.data);
-        expect(called).toBe(id);
-        expect(map.hasImage(id)).toBeTruthy();
-    });
-
-    test('map getImage matches addImage, uintArray', () => {
-        const map = createMap();
-        const id = 'add-get-uint';
-        const inputImage = {width: 2, height: 1, data: new Uint8Array(8)};
-
-        map.addImage(id, inputImage);
-        expect(map.hasImage(id)).toBeTruthy();
-
-        const gotImage = map.getImage(id);
-        expect(gotImage.data.width).toEqual(inputImage.width);
-        expect(gotImage.data.height).toEqual(inputImage.height);
-        expect(gotImage.sdf).toBe(false);
-    });
-
-    test('map getImage matches addImage, uintClampedArray', () => {
-        const map = createMap();
-        const id = 'add-get-uint-clamped';
-        const inputImage = {width: 1, height: 2, data: new Uint8ClampedArray(8)};
-
-        map.addImage(id, inputImage);
-        expect(map.hasImage(id)).toBeTruthy();
-
-        const gotImage = map.getImage(id);
-        expect(gotImage.data.width).toEqual(inputImage.width);
-        expect(gotImage.data.height).toEqual(inputImage.height);
-        expect(gotImage.sdf).toBe(false);
-    });
-
-    test('map getImage matches addImage, ImageData', () => {
-        const map = createMap();
-        const id = 'add-get-image-data';
-        const inputImage = new ImageData(1, 3);
-
-        map.addImage(id, inputImage);
-        expect(map.hasImage(id)).toBeTruthy();
-
-        const gotImage = map.getImage(id);
-        expect(gotImage.data.width).toEqual(inputImage.width);
-        expect(gotImage.data.height).toEqual(inputImage.height);
-        expect(gotImage.sdf).toBe(false);
-    });
-
-    test('map getImage matches addImage, StyleImageInterface uint', () => {
-        const map = createMap();
-        const id = 'add-get-style-image-iface-uint';
-        const inputImage: StyleImageInterface = {
-            width: 3,
-            height: 1,
-            data: new Uint8Array(12)
-        };
-
-        map.addImage(id, inputImage);
-        expect(map.hasImage(id)).toBeTruthy();
-
-        const gotImage = map.getImage(id);
-        expect(gotImage.data.width).toEqual(inputImage.width);
-        expect(gotImage.data.height).toEqual(inputImage.height);
-        expect(gotImage.sdf).toBe(false);
-    });
-
-    test('map getImage matches addImage, StyleImageInterface clamped', () => {
-        const map = createMap();
-        const id = 'add-get-style-image-iface-clamped';
-        const inputImage: StyleImageInterface = {
-            width: 4,
-            height: 1,
-            data: new Uint8ClampedArray(16)
-        };
-
-        map.addImage(id, inputImage);
-        expect(map.hasImage(id)).toBeTruthy();
-
-        const gotImage = map.getImage(id);
-        expect(gotImage.data.width).toEqual(inputImage.width);
-        expect(gotImage.data.height).toEqual(inputImage.height);
-        expect(gotImage.sdf).toBe(false);
-    });
-
-    test('map getImage matches addImage, StyleImageInterface SDF', () => {
-        const map = createMap();
-        const id = 'add-get-style-image-iface-sdf';
-        const inputImage: StyleImageInterface = {
-            width: 5,
-            height: 1,
-            data: new Uint8Array(20)
-        };
-
-        map.addImage(id, inputImage, {sdf: true});
-        expect(map.hasImage(id)).toBeTruthy();
-
-        const gotImage = map.getImage(id);
-        expect(gotImage.data.width).toEqual(inputImage.width);
-        expect(gotImage.data.height).toEqual(inputImage.height);
-        expect(gotImage.sdf).toBe(true);
-    });
-
-    test('map does not fire `styleimagemissing` for empty icon values', done => {
-        const map = createMap();
-
-        map.on('load', () => {
-            map.on('idle', () => {
-                done();
-            });
-
-            map.addSource('foo', {
-                type: 'geojson',
-                data: {type: 'Point', coordinates: [0, 0]}
-            });
-            map.addLayer({
-                id: 'foo',
-                type: 'symbol',
-                source: 'foo',
-                layout: {
-                    'icon-image': ['case', true, '', '']
-                }
-            });
-
-            map.on('styleimagemissing', ({id}) => {
-                done(`styleimagemissing fired for value ${id}`);
-            });
-        });
-    });
-
-    describe('setPixelRatio', () => {
-        test('resizes canvas', () => {
-            const container = window.document.createElement('div');
-            Object.defineProperty(container, 'clientWidth', {value: 512});
-            Object.defineProperty(container, 'clientHeight', {value: 512});
-            const map = createMap({container, pixelRatio: 1});
-            expect(map.getCanvas().width).toBe(512);
-            expect(map.getCanvas().height).toBe(512);
-            map.setPixelRatio(2);
-            expect(map.getCanvas().width).toBe(1024);
-            expect(map.getCanvas().height).toBe(1024);
-        });
-
-        test('resizes painter', () => {
-            const container = window.document.createElement('div');
-            Object.defineProperty(container, 'clientWidth', {value: 512});
-            Object.defineProperty(container, 'clientHeight', {value: 512});
-            const map = createMap({container, pixelRatio: 1});
-            expect(map.painter.pixelRatio).toBe(1);
-            expect(map.painter.width).toBe(512);
-            expect(map.painter.height).toBe(512);
-            map.setPixelRatio(2);
-            expect(map.painter.pixelRatio).toBe(2);
-            expect(map.painter.width).toBe(1024);
-            expect(map.painter.height).toBe(1024);
-        });
-    });
-
-    describe('getPixelRatio', () => {
-        test('returns the pixel ratio', () => {
-            const map = createMap({pixelRatio: 1});
-            expect(map.getPixelRatio()).toBe(1);
-            map.setPixelRatio(2);
-            expect(map.getPixelRatio()).toBe(2);
-        });
-    });
-
-    test('pixel ratio defaults to devicePixelRatio', () => {
-        const map = createMap();
-        expect(map.getPixelRatio()).toBe(devicePixelRatio);
-    });
-
-    test('pixel ratio by default reflects devicePixelRatio changes', () => {
-        global.devicePixelRatio = 0.25;
-        const map = createMap();
-        expect(map.getPixelRatio()).toBe(0.25);
-        global.devicePixelRatio = 1;
-        expect(map.getPixelRatio()).toBe(1);
-    });
-
-    test('canvas has the expected size', () => {
-        const container = window.document.createElement('div');
-        Object.defineProperty(container, 'clientWidth', {value: 512});
-        Object.defineProperty(container, 'clientHeight', {value: 512});
-        const map = createMap({container, pixelRatio: 2});
-        expect(map.getCanvas().width).toBe(1024);
-        expect(map.getCanvas().height).toBe(1024);
-    });
-
-    test('painter has the expected size and pixel ratio', () => {
-        const container = window.document.createElement('div');
-        Object.defineProperty(container, 'clientWidth', {value: 512});
-        Object.defineProperty(container, 'clientHeight', {value: 512});
-        const map = createMap({container, pixelRatio: 2});
-        expect(map.painter.pixelRatio).toBe(2);
-        expect(map.painter.width).toBe(1024);
-        expect(map.painter.height).toBe(1024);
-    });
-
     test('fires sourcedataabort event on dataabort event', async () => {
         const map = createMap();
         const sourcePromise = map.once('sourcedataabort');
         map.fire(new EventedEvent('dataabort'));
         await sourcePromise;
-    });
-
-    describe('#setTerrain', () => {
-        test('warn when terrain and hillshade source identical', done => {
-            server.respondWith('/source.json', JSON.stringify({
-                minzoom: 5,
-                maxzoom: 12,
-                attribution: 'Terrain',
-                tiles: ['http://example.com/{z}/{x}/{y}.pngraw'],
-                bounds: [-47, -7, -45, -5]
-            }));
-
-            const map = createMap();
-
-            map.on('load', () => {
-                map.addSource('terrainrgb', {type: 'raster-dem', url: '/source.json'});
-                server.respond();
-                map.addLayer({id: 'hillshade', type: 'hillshade', source: 'terrainrgb'});
-                const stub = jest.spyOn(console, 'warn').mockImplementation(() => { });
-                stub.mockReset();
-                map.setTerrain({
-                    source: 'terrainrgb'
-                });
-                expect(console.warn).toHaveBeenCalledTimes(1);
-                done();
-            });
-        });
-    });
-
-    describe('#getTerrain', () => {
-        test('returns null when not set', () => {
-            const map = createMap();
-            expect(map.getTerrain()).toBeNull();
-        });
     });
 
     describe('cooperativeGestures option', () => {
@@ -1473,26 +569,6 @@ describe('Map', () => {
             const map = createMap({cooperativeGestures: true});
             map.cooperativeGestures.disable();
             expect(map.getContainer().querySelector('.maplibregl-cooperative-gesture-screen')).toBeFalsy();
-        });
-    });
-
-    describe('getCameraTargetElevation', () => {
-        test('Elevation is zero without terrain, and matches any given terrain', () => {
-            const map = createMap();
-            expect(map.getCameraTargetElevation()).toBe(0);
-
-            const terrainStub = {} as Terrain;
-            map.terrain = terrainStub;
-
-            const transform = new Transform(0, 22, 0, 60, true);
-            transform.elevation = 200;
-            transform.center = new LngLat(10.0, 50.0);
-            transform.zoom = 14;
-            transform.resize(512, 512);
-            transform.elevation = 2000;
-            map.transform = transform;
-
-            expect(map.getCameraTargetElevation()).toBe(2000);
         });
     });
 
@@ -1597,99 +673,4 @@ describe('Map', () => {
             expect(mockedGetElevation.mock.calls).toHaveLength(0);
         });
     });
-
-    describe('webgl errors', () => {
-        test('WebGL error while creating map', () => {
-            const original = HTMLCanvasElement.prototype.getContext;
-            HTMLCanvasElement.prototype.getContext = function (type: string) {
-                if (type === 'webgl2' || type === 'webgl') {
-                    const errorEvent = new Event('webglcontextcreationerror');
-                    (errorEvent as any).statusMessage = 'mocked webglcontextcreationerror message';
-                    (this as HTMLCanvasElement).dispatchEvent(errorEvent);
-                    return null;
-                }
-            };
-            try {
-                createMap();
-            } catch (e) {
-                const errorMessageObject = JSON.parse(e.message);
-
-                // this message is from map code
-                expect(errorMessageObject.message).toBe('Failed to initialize WebGL');
-
-                // this is from test mock
-                expect(errorMessageObject.statusMessage).toBe('mocked webglcontextcreationerror message');
-            } finally {
-                HTMLCanvasElement.prototype.getContext = original;
-            }
-        });
-        test('Hit WebGL max drawing buffer limit', () => {
-            // Simulate a device with MAX_TEXTURE_SIZE=16834 and max rendering area of ~32Mpx
-            const container = window.document.createElement('div');
-            Object.defineProperty(container, 'clientWidth', {value: 8000});
-            Object.defineProperty(container, 'clientHeight', {value: 4500});
-            const map = createMap({container, maxCanvasSize: [16834, 16834], pixelRatio: 1});
-            jest.spyOn(map.painter.context.gl, 'drawingBufferWidth', 'get').mockReturnValue(7536);
-            jest.spyOn(map.painter.context.gl, 'drawingBufferHeight', 'get').mockReturnValue(4239);
-            map.resize();
-            expect(map.getCanvas().width).toBe(7536);
-            expect(map.getCanvas().height).toBe(4239);
-            // Check if maxCanvasSize is updated
-            expect(map._maxCanvasSize).toEqual([7536, 4239]);
-        });
-    });
-
-    describe('Max Canvas Size option', () => {
-        test('maxCanvasSize width = height', () => {
-            const container = window.document.createElement('div');
-            Object.defineProperty(container, 'clientWidth', {value: 2048});
-            Object.defineProperty(container, 'clientHeight', {value: 2048});
-            jest.spyOn(WebGLRenderingContext.prototype, 'drawingBufferWidth', 'get').mockReturnValue(8192);
-            jest.spyOn(WebGLRenderingContext.prototype, 'drawingBufferHeight', 'get').mockReturnValue(8192);
-            const map = createMap({container, maxCanvasSize: [8192, 8192], pixelRatio: 5});
-            map.resize();
-            expect(map.getCanvas().width).toBe(8192);
-            expect(map.getCanvas().height).toBe(8192);
-        });
-
-        test('maxCanvasSize width != height', () => {
-            const container = window.document.createElement('div');
-            Object.defineProperty(container, 'clientWidth', {value: 1024});
-            Object.defineProperty(container, 'clientHeight', {value: 2048});
-            jest.spyOn(WebGLRenderingContext.prototype, 'drawingBufferWidth', 'get').mockReturnValue(8192);
-            jest.spyOn(WebGLRenderingContext.prototype, 'drawingBufferHeight', 'get').mockReturnValue(4096);
-            const map = createMap({container, maxCanvasSize: [8192, 4096], pixelRatio: 3});
-            map.resize();
-            expect(map.getCanvas().width).toBe(2048);
-            expect(map.getCanvas().height).toBe(4096);
-        });
-
-        test('maxCanvasSize below clientWidth and clientHeigth', () => {
-            const container = window.document.createElement('div');
-            Object.defineProperty(container, 'clientWidth', {value: 12834});
-            Object.defineProperty(container, 'clientHeight', {value: 9000});
-            jest.spyOn(WebGLRenderingContext.prototype, 'drawingBufferWidth', 'get').mockReturnValue(4096);
-            jest.spyOn(WebGLRenderingContext.prototype, 'drawingBufferHeight', 'get').mockReturnValue(8192);
-            const map = createMap({container, maxCanvasSize: [4096, 8192], pixelRatio: 1});
-            map.resize();
-            expect(map.getCanvas().width).toBe(4096);
-            expect(map.getCanvas().height).toBe(2872);
-        });
-
-        test('maxCanvasSize with setPixelRatio', () => {
-            const container = window.document.createElement('div');
-            Object.defineProperty(container, 'clientWidth', {value: 2048});
-            Object.defineProperty(container, 'clientHeight', {value: 2048});
-            jest.spyOn(WebGLRenderingContext.prototype, 'drawingBufferWidth', 'get').mockReturnValue(3072);
-            jest.spyOn(WebGLRenderingContext.prototype, 'drawingBufferHeight', 'get').mockReturnValue(3072);
-            const map = createMap({container, maxCanvasSize: [3072, 3072], pixelRatio: 1.25});
-            map.resize();
-            expect(map.getCanvas().width).toBe(2560);
-            expect(map.getCanvas().height).toBe(2560);
-            map.setPixelRatio(2);
-            expect(map.getCanvas().width).toBe(3072);
-            expect(map.getCanvas().height).toBe(3072);
-        });
-    });
-
 });
